Give low Z covariance in blob detector and nearest neighbor tracker
The default was to give very high (mimicing "infinite") covariance for the Z
position of both detections and tracks since we're only tracking in two
dimensions.  But this causes the covariance ovals visualized in Rviz to be
enormous.

For our use case, it's not that we don't know the Z position, it's that it
never changes since people, carts, etc don't fly, so a low covariance is a
better reflection of reality.