Reduce the amount of data being logged by moving a majority to DEBUG
We were seeing syslogs with sizes pushing 100GBs because of the history
was being dumped multiple times per second and each history was 100+ lines
long!

To balance data with verbosity the same data will still be logged but
only if this code is built in "Debug" mode (see the CMakeLists.txt file
for what that means).