messages: removed unneeded dependencies
Messages packages don't need roscpp or roslib since they just need to be
able to generate the messages. The actual code packages that use these
has dependencies on them and will cause them to be pulled in as needed.
When building roscpp has been added to the catkin workspace rebuilding
just the message packages were blocked by waiting on roscpp, which would
significantly increase the amount of time needed just to rebuild the
message packages.